Update setup.md
Added link for The Biostars Handbook. 

Also I followed along with the instructions on a Mac instead of an Amazon instance, since all of the packages indicated they are available for Mac. This went fine except for BCFtools, which failed to compile. I didn't chase the problem, but I'm guessing a missing library.

I then tried installing homebrew and miniconda to see if they could successfully install BCFtools. Homebrew failed to install based on the command provided at https://brew.sh/. The miniconda install was successful, but then was unable to install BCFtools because it couldn't find BCFtools in the channels it searched: 

PackageNotFoundError: Packages missing in current channels:
            
  - bcftools

We have searched for the packages in the following channels:
            
  - https://repo.continuum.io/pkgs/main/osx-64
  - https://repo.continuum.io/pkgs/main/noarch
  - https://repo.continuum.io/pkgs/free/osx-64
  - https://repo.continuum.io/pkgs/free/noarch
  - https://repo.continuum.io/pkgs/r/osx-64
  - https://repo.continuum.io/pkgs/r/noarch
  - https://repo.continuum.io/pkgs/pro/osx-64
  - https://repo.continuum.io/pkgs/pro/noarch

I'd recommend either making it clear that the Amazon environment is required and that the steps are not guaranteed to work on other platforms, or update and test the instructions on Mac and PC as well.