remove indentation call out
Bash for loop _can_ be indented, but they do not  "need to be indented" as stated in this callout. indenting is useful for humans, but it is not necessary. 

This is not the instance of a for loop in this lesson, so I think this is too late for a callout here. One could be added earlier, but I would suggest including something about one line versus for line for loops.