Add comma for readability
I tripped over this sentence while reading it: "3) you will often need to customize your analyses and command-line tools often enable". I had to re-read it a few times before I realized that the "and" is the start of a new sentence. I am proposing a comma before the "and" because I think the comma increases readability.

After a bit of digging, I think I have figured out the source of my confusion. Sentence 3) is a compound sentence bound together by "and", and such sentences take a comma before the "and". Incidentally, the previous sentence is a compound sentence too. For reference, see rule /issues/3 on [here](https://webapps.towson.edu/ows/modulecomma.htm)